fix: extend retry function to mutate rules
Uses the UpdateRetryAnnotation mechanism already implemented in generate rules for mutate rules.

This fixes the issue that UpdateRequests keep existing indefinitely when the original trigger resource was deleted.
